refactor: deprecate packages/effect, move Terrain Grass into packages/geometry (#446)
