+++ package.json
@@ -32,13 +32,13 @@
         "dev": "vite",
         "build": "tsc --p tsconfig.build.json && vite build && npm run build:types && npm run minify",
         "build:test": "tsc --p tsconfig.build.json && vite build && npm run build:packages",
-        "build:packages": "npm run build:physics && npm run build:media && npm run build:stats && npm run build:particle && npm run build:graphic && npm run build:effect",
+        "build:packages": "npm run build:physics && npm run build:media && npm run build:stats && npm run build:particle && npm run build:graphic && npm run build:geometry",
         "build:physics": "cd packages/physics && npm run build",
         "build:media": "cd packages/media-extention && npm run build",
         "build:stats": "cd packages/stats && npm run build",
         "build:particle": "cd packages/particle && npm run build",
         "build:graphic": "cd packages/graphic && npm run build",
-        "build:effect": "cd packages/effect && npm run build",
+        "build:geometry": "cd packages/geometry && npm run build",
         "build:types": "tsc --emitDeclarationOnly -p tsconfig.build.json && rm -rf dist/packages && mv dist/src dist/types",
         "minify": "./node_modules/vite/node_modules/.bin/esbuild dist/orillusion.es.max.js --sourcemap --minify --outfile=dist/orillusion.es.js && ./node_modules/vite/node_modules/.bin/esbuild dist/orillusion.umd.max.js --minify --sourcemap --outfile=dist/orillusion.umd.js",
         "test": "electron test/ci/main.js",
