--- packages/effect/README.md
@@ -1,15 +1,0 @@
-Add effect components for [Orillusion](https://www.orillusion.com)
-
-## Install
-```bash
-npm install @orillusion/core --save
-npm install @orillusion/effect --save
-```
-
-Or access Global build from CDN
-```html
-<script src="https://unpkg.com/@orillusion/core/dist/orillusion.umd.js"></script>
-<script src="https://unpkg.com/@orillusion/media-extention/dist/effect.umd.js"></script>
-```
-
-More doc from [Orillusion](https://www.orillusion.com/guide/effect/Readme.html)
