--- packages/effect/grass/GrassNode.ts
@@ -1,5 +1,0 @@
-import { Transform } from "@orillusion/core";
-
-export class GrassNode extends Transform {
-
-}
