--- packages/effect/index.ts
@@ -1,5 +1,0 @@
-export * from "./grass/GrassNode";
-export * from "./grass/component/GrassComponent";
-export * from "./grass/geometry/GrassGeometry";
-export * from "./grass/material/GrassMaterial";
-export * from "./terrain/geometry/TerrainGeometry";
