--- packages/effect/package.json
@@ -1,24 +1,0 @@
-{
-    "name": "@orillusion/effect",
-    "version": "0.1.3",
-    "author": "Orillusion",
-    "description": "Orillusion Effects Plugin",
-    "main": "./dist/effect.umd.js",
-    "module": "./dist/effect.es.js",
-    "module:dev": "./index.ts",
-    "types": "./dist/index.d.ts",
-    "files": ["dist"],
-    "scripts": {
-        "build": "vite build && npm run build:types && npm run build:clean",
-        "build:types": "tsc --emitDeclarationOnly -p tsconfig.json",
-        "build:clean": "mv dist/packages/effect/* dist && rm -rf dist/src && rm -rf dist/packages"
-    },
-    "license": "MIT",
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/Orillusion/orillusion.git"
-    },
-    "dependencies": {
-        "@orillusion/core": "^0.8.0"
-    }
-}
