+++ packages/geometry/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@orillusion/geometry",
-    "version": "0.2.1",
+    "version": "0.2.2",
     "author": "Orillusion",
     "description": "Orillusion geometry Plugin",
     "main": "./dist/geometry.umd.js",
